+++ anathema-value-resolver/src/functions/mod.rs
@@ -57,6 +57,7 @@ impl FunctionTable {
         let mut inner = HashMap::new();
         inner.insert("to_upper".into(), Function::from(string::to_upper));
         inner.insert("to_lower".into(), Function::from(string::to_lower));
+        inner.insert("truncate".into(), Function::from(string::truncate));
         inner.insert("to_str".into(), Function::from(string::to_str));
         inner.insert("to_int".into(), Function::from(number::to_int));
         inner.insert("to_float".into(), Function::from(number::to_float));