+++ anathema-value-resolver/src/functions/string.rs
@@ -1,3 +1,5 @@
+use unicode_width::{UnicodeWidthChar, UnicodeWidthStr};
+
 use crate::ValueKind;
 
 pub(super) fn to_upper<'bp>(args: &[ValueKind<'bp>]) -> ValueKind<'bp> {
@@ -35,6 +37,37 @@ pub(super) fn to_str<'bp>(args: &[ValueKind<'bp>]) -> ValueKind<'bp> {
     let mut buffer = String::new();
     args[0].strings(|s| {
         buffer.push_str(s);
+        true
+    });
+    ValueKind::Str(buffer.into())
+}
+
+pub(super) fn truncate<'bp>(args: &[ValueKind<'bp>]) -> ValueKind<'bp> {
+    if args.len() != 2 {
+        return ValueKind::Null;
+    }
+
+    let Some(mut remainder) = args[1].as_int() else { return ValueKind::Null };
+
+    let mut buffer = String::new();
+    args[0].strings(|s| {
+        let width = s.width() as i64;
+
+        if width < remainder {
+            remainder = remainder.saturating_sub(width);
+            buffer.push_str(s);
+        } else {
+            let mut chars = s.chars();
+            while remainder > 0 {
+                let Some(c) = chars.next() else { break };
+                let width = c.width().unwrap_or(0) as i64;
+                if width > remainder {
+                    break;
+                };
+                buffer.push(c);
+                remainder = remainder.saturating_sub(width);
+            }
+        }
         true
     });
     ValueKind::Str(buffer.into())
@@ -85,5 +118,26 @@ mod test {
         let val = list([1, 2, 3]);
         let val = to_str(&[val]);
         assert_eq!(val, value("123"));
+    }
+
+    #[test]
+    fn truncate_long_string() {
+        let val = value("this is a longer string");
+        let val = truncate(&[val, value(4)]);
+        assert_eq!(val, value("this"));
+    }
+
+    #[test]
+    fn truncate_short_string() {
+        let val = value("hi");
+        let val = truncate(&[val, value(4)]);
+        assert_eq!(val, value("hi"));
+    }
+
+    #[test]
+    fn truncate_wide_cells() {
+        let val = value("🐇🐇");
+        let val = truncate(&[val, value(3)]);
+        assert_eq!(val, value("🐇"));
     }
 }